create hook for Sign Up states
