+++ src/screens/SignUpScreen/index.js
@@ -1,4 +1,4 @@
-import React, { useCallback, useState } from 'react';
+import React, { useCallback } from 'react';
 import {
   View,
   SafeAreaView,
@@ -7,85 +7,39 @@ import {
   ImageBackground,
 } from 'react-native';
 import Header from 'components/common/Header';
-import Input from 'components/common/Form/Input';
-import Button from 'components/common/Form/Button';
+import Input from 'components/common/form/Input';
+import Button from 'components/common/form/Button';
 import styles from './styles';
 import strings from 'locale';
+import loginIcon from 'assets/logoLogin.png';
+import useSignUpStates from 'hooks/useSignUpStates';
+
+const { SIGN_UP, SIGN_UP_HELP } = strings;
 
 const SignUp = ({ navigation }) => {
   const signIn = useCallback(() => navigation.goBack());
 
-  const [name, setName] = useState('');
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [confirmPassword, setConfirmPassword] = useState('');
-  const [gender, setGender] = useState('');
-
-  const [errorName, setErrorName] = useState('');
-  const [errorEmail, setErrorEmail] = useState('');
-  const [errorPassword, setErrorPassword] = useState('');
-  const [errorConfirmPassword, setErrorConfirmPassword] = useState('');
-  const [errorGender, setErrorGender] = useState('');
-
-  const { SIGN_UP, SIGN_UP_ERROR, SIGN_UP_HELP } = strings;
-
-  const cbName = newName => {
-    setName(newName);
-    setErrorName('');
-  };
-
-  const cbEmail = newEmail => {
-    setEmail(newEmail);
-    setErrorEmail('');
-  };
-
-  const cbPassword = newPassword => {
-    setPassword(newPassword);
-    setErrorPassword('');
-  };
-
-  const cbConfirmPassword = newConfirmPassword => {
-    setConfirmPassword(newConfirmPassword);
-    setErrorConfirmPassword('');
-  };
-
-  const cbGender = newGender => {
-    setGender(newGender);
-    setErrorGender('');
-  };
-
-  const cbSignUp = () => {
-    setErrorEmail('');
-    setErrorEmail('');
-    setErrorPassword('');
-    setErrorConfirmPassword('');
-    setErrorGender('');
-
-    if (!name) {
-      setErrorName(SIGN_UP_ERROR.emptyName);
-    }
-
-    if (!email || !email.includes('@')) {
-      setErrorEmail(SIGN_UP_ERROR.emptyEmail);
-    }
-
-    if (password.length < 6) {
-      setErrorPassword(SIGN_UP_ERROR.passwordSixChar);
-    }
-
-    if (confirmPassword !== password) {
-      setErrorConfirmPassword(SIGN_UP_ERROR.confirmPasswordMatch);
-    }
-
-    if (!gender) {
-      setErrorGender(SIGN_UP_ERROR.emptyGender);
-    }
-  };
+  const {
+    name,
+    email,
+    password,
+    confirmPassword,
+    gender,
+    errorName,
+    errorEmail,
+    errorPassword,
+    errorConfirmPassword,
+    errorGender,
+    cbName,
+    cbEmail,
+    cbPassword,
+    cbConfirmPassword,
+    cbGender,
+    cbSignUp,
+  } = useSignUpStates();
 
   return (
-    <ImageBackground
-      source={require('assets/logoLogin.png')}
-      style={styles.image}>
+    <ImageBackground source={loginIcon} style={styles.image}>
       <Header title={strings.COMMON.headerTitle} />
       <SafeAreaView style={styles.safeArea}>
         <Input
